+++ Rhombus/src/Rhombus/ECS/Components/TweenComponent.h
@@ -21,7 +21,7 @@ namespace rhombus
 		Ref<Tween> CreateTween(Vec4* param, Vec4 begin, Vec4 finish, float duration, EasingType easingType = EasingType::LINEAR) { m_tweens.push_back(std::make_shared<Tween>(param, begin, finish, duration, easingType)); return m_tweens.back(); }
 
 		Ref<Tween> GetTween(int i) { return m_tweens[i]; }
-		int GetTweenCount() const { return m_tweens.size(); }
+		int GetTweenCount() const { return (int)m_tweens.size(); }
 
 	private:
 		std::vector<Ref<Tween>> m_tweens;
